+++ tests/conftest.py
@@ -1,9 +1,15 @@
 import os
 
+import hipscat as hc
+import lsdb
 import pytest
 import shortuuid
+from hipscat.io.file_io import file_io
 
 from hipscat_cloudtests.temp_cloud_directory import TempCloudDirectory
+
+SMALL_SKY_XMATCH_NAME = "small_sky_xmatch"
+XMATCH_CORRECT_FILE = "xmatch_correct.csv"
 
 ALMANAC_DIR_NAME = "almanac"
 SMALL_SKY_DIR_NAME = "small_sky"
@@ -113,3 +119,40 @@ def tmp_cloud_path(request, tmp_dir_cloud):
     # Strip out the "test_" at the beginning of each method name, make it a little
     # shorter, and add a disambuating UUID.
     return f"{tmp_dir_cloud}/{name[5:25]}_{my_uuid}"
+
+
+@pytest.fixture
+def small_sky_xmatch_dir_cloud(cloud_path):
+    return os.path.join(cloud_path, "data", SMALL_SKY_XMATCH_NAME)
+
+
+@pytest.fixture
+def small_sky_catalog_cloud(small_sky_dir_cloud, storage_options):
+    return lsdb.read_hipscat(small_sky_dir_cloud, storage_options=storage_options)
+
+
+@pytest.fixture
+def small_sky_xmatch_catalog_cloud(small_sky_xmatch_dir_cloud, storage_options):
+    return lsdb.read_hipscat(small_sky_xmatch_dir_cloud, storage_options=storage_options)
+
+
+@pytest.fixture
+def small_sky_order1_hipscat_catalog_cloud(small_sky_order1_dir_cloud, storage_options):
+    return hc.catalog.Catalog.read_from_hipscat(small_sky_order1_dir_cloud, storage_options=storage_options)
+
+
+@pytest.fixture
+def small_sky_order1_catalog_cloud(small_sky_order1_dir_cloud, storage_options):
+    return lsdb.read_hipscat(small_sky_order1_dir_cloud, storage_options=storage_options)
+
+
+@pytest.fixture
+def xmatch_correct_cloud(local_data_dir):
+    pathway = os.path.join(local_data_dir, "xmatch", XMATCH_CORRECT_FILE)
+    return file_io.load_csv_to_pandas(pathway)
+
+
+@pytest.fixture
+def xmatch_with_margin(local_data_dir):
+    pathway = os.path.join(local_data_dir, "xmatch", "xmatch_with_margin.csv")
+    return file_io.load_csv_to_pandas(pathway)
